LayoutTests/fast/css/translate-invalidate-if-disabled.html

<!DOCTYPE html><!-- webkit-test-runner [ CSSIndividualTransformPropertiesEnabled=false ] -->
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="target" style="display: none"></div>
<script>
description("Tests that translate is not exposed when the feature is disabled");

shouldBeFalse("'translate' in document.documentElement.style");
shouldBeFalse("'translate' in getComputedStyle(document.documentElement)");
shouldBeTrue("getComputedStyle(document.getElementById('target')).getPropertyValue('translate') == ''");
shouldBeFalse("CSS.supports('translate: contain')");
shouldBeFalse("CSS.supports('translate: inherit')");
shouldBeFalse("CSS.supports('translate', 'inherit')");

</script>
</body>
</html>
