LayoutTests/fast/css/translate-invalidate-if-disabled-expected.txt

Tests that translate is not exposed when the feature is disabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'translate' in document.documentElement.style is false
PASS 'translate' in getComputedStyle(document.documentElement) is false
PASS getComputedStyle(document.getElementById('target')).getPropertyValue('translate') == '' is true
PASS CSS.supports('translate: contain') is false
PASS CSS.supports('translate: inherit') is false
PASS CSS.supports('translate', 'inherit') is false
PASS successfullyParsed is true

TEST COMPLETE
